Fix dialog controllers' references
